Split of model loader and context mostly similar to how llama.cpp did that already: falcon_context now contains the falcon_model and a pointer to vocab the model contains the vocab now this allows to now conveniently load more than one context as well as loading multiple models at once
A ton of changes - if anything acts funky please report

+++ examples/falcon_common.cpp
@@ -384,7 +384,7 @@ bool gpt_params_parse(int argc, char ** argv, gpt_params & params) {
             }
             #ifdef GGML_USE_CUBLAS
             params.mb_reserve_gpu_main = std::stoi(argv[i]);
-            ggml_cuda_set_vram_reserved(params.mb_reserve_gpu_main * 1024*1024);
+            ggml_cuda_set_vram_reserved(((int64_t)params.mb_reserve_gpu_main)*1024*1024);
             #else
             fprintf(stderr, "warning: falcon.cpp was compiled without cuBLAS. VRAM not available.\n");
             #endif
